pyproject.toml
[project]
name = "dunder-mifflin-play"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.3.0",
    "google-cloud-aiplatform[adk,agent-engines]>=1.97.0",
    "google-genai>=1.20.0",
    "httpx>=0.28.1",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.7.0",
]
